Realization/Neural network/Realization/Neural Network/requirements.txt
torch
torchvision
torchaudio
pandas
numpy
pyyaml
tqdm
scikit-learn
regex
matplotlib